split model list serializer
