+++ codeforlife/serializers/__init__.py
@@ -4,4 +4,5 @@ Created on 20/01/2024 at 11:19:12(+00:00).
 """
 
 from .base import BaseSerializer
-from .model import ModelListSerializer, ModelSerializer
+from .model import ModelSerializer
+from .model_list import ModelListSerializer
